Basic protocol association and font preview window
Very unfinished. Need to properly manage Window lifecycles.
