+++ CharacterMap/CharacterMap/App.xaml
@@ -16,6 +16,8 @@
                     <converters1:ZoomBackgroundConverter x:Key="ZoomBackgroundConverter" />
                 </ResourceDictionary>
                 <ResourceDictionary>
+                    <x:Boolean x:Key="TrueValue">True</x:Boolean>
+                    <x:Boolean x:Key="FalseValue">False</x:Boolean>
                     <core:AppSettings x:Key="AppSettings" />
                     <Style x:Key="CharItemGridStyle" TargetType="Grid">
                         <!--<Setter Property="BorderThickness" Value="1" />
