+++ CharacterMap/CharacterMap/Package.appxmanifest
@@ -25,6 +25,19 @@
         </uap:DefaultTile>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
       </uap:VisualElements>
+      <Extensions>
+        <uap:Extension Category="windows.fileTypeAssociation">
+          <uap:FileTypeAssociation Name="fontfile">
+            <uap:SupportedFileTypes>
+              <!--<uap:FileType ContentType="application/x-font-ttf">.ttf</uap:FileType>-->
+              <uap:FileType>.otf</uap:FileType>
+            </uap:SupportedFileTypes>
+            <uap:DisplayName>Character Map</uap:DisplayName>
+            <uap:InfoTip>Preview Font</uap:InfoTip>
+            <uap:EditFlags OpenIsSafe="true"/>
+          </uap:FileTypeAssociation>
+        </uap:Extension>
+      </Extensions>
     </Application>
   </Applications>
   <Capabilities>
