+++ CharacterMap/CharacterMap/ViewModels/MainViewModel.cs
@@ -35,33 +35,16 @@ namespace CharacterMap.ViewModels
 
     public class MainViewModel : ViewModelBase
     {
-        private IReadOnlyList<Character> _chars;
-        private string _fontIcon;
         private ObservableCollection<AlphaKeyGroup<InstalledFont>> _groupedFontList;
-        private Character _selectedChar;
         private InstalledFont _selectedFont;
-        private string _xamlCode;
-        private string _symbolIcon;
 
-        private Interop _interop { get; }
-
-        public ExportStyle BlackColor { get; } = ExportStyle.Black;
-        public ExportStyle WhiteColor { get; } = ExportStyle.White;
-        public ExportStyle GlyphColor { get; } = ExportStyle.ColorGlyph;
-
+        public IDialogService DialogService { get; }
 
         public MainViewModel(IDialogService dialogService)
         {
             DialogService = dialogService;
             AppNameVersion = GetAppDescription();
-            CommandSavePng = new RelayCommand<ExportStyle>(async (b) => await SavePngAsync(b));
-            CommandSaveSvg = new RelayCommand<bool>(async (b) => await SaveSvgAsync(b));
             CommandToggleFullScreen = new RelayCommand(ToggleFullScreenMode);
-
-            SimpleIoc.Default.Register(() => new Interop(CanvasDevice.GetSharedDevice()));
-            _interop = SimpleIoc.Default.GetInstance<Interop>();
-
-
             Load();
         }
 
@@ -98,38 +81,11 @@ namespace CharacterMap.ViewModels
             set => Set(ref _titlePrefix, value);
         }
 
-        public IDialogService DialogService { get; set; }
-
-        public RelayCommand<ExportStyle> CommandSavePng { get; set; }
-
-        public RelayCommand<bool> CommandSaveSvg { get; set; }
-
-        private FontVariant _selectedVariant;
-        public FontVariant SelectedVariant
-        {
-            get => _selectedVariant;
-            set { if (Set(ref _selectedVariant, value)) LoadChars(value); }
-        }
-
-        private bool _showColorGlyphs = true;
-        public bool ShowColorGlyphs
-        {
-            get => _showColorGlyphs;
-            set => Set(ref _showColorGlyphs, value);
-        }
-
         private bool _isLoadingFonts;
         public bool IsLoadingFonts
         {
             get => _isLoadingFonts;
             set => Set(ref _isLoadingFonts, value);
-        }
-
-        private CanvasTextLayoutAnalysis _selectedCharAnalysis;
-        public CanvasTextLayoutAnalysis SelectedCharAnalysis
-        {
-            get => _selectedCharAnalysis;
-            set => Set(ref _selectedCharAnalysis, value);
         }
 
         private List<InstalledFont> _fontList;
@@ -151,51 +107,6 @@ namespace CharacterMap.ViewModels
         {
             get => _groupedFontList;
             set => Set(ref _groupedFontList, value);
-        }
-
-        public IReadOnlyList<Character> Chars
-        {
-            get => _chars;
-            set => Set(ref _chars, value);
-        }
-
-        public Character SelectedChar
-        {
-            get => _selectedChar;
-            set
-            {
-                _selectedChar = value;
-                if (null != value)
-                {
-                    XamlCode = $"&#x{value.UnicodeIndex.ToString("x").ToUpper()};";
-                    FontIcon = $@"<FontIcon FontFamily=""{SelectedFont.Name}"" Glyph=""&#x{
-                            value.UnicodeIndex.ToString("x").ToUpper()
-                        };"" />";
-                    SymbolIcon = $"(Symbol)0x{value.UnicodeIndex.ToString("x").ToUpper()}";
-
-                    App.AppSettings.LastSelectedCharIndex = value.UnicodeIndex;
-                }
-                RaisePropertyChanged();
-                UpdateCharAnalysis();
-            }
-        }
-
-        public string XamlCode
-        {
-            get => _xamlCode;
-            set => Set(ref _xamlCode, value);
-        }
-
-        public string SymbolIcon
-        {
-            get => _symbolIcon;
-            set => Set(ref _symbolIcon, value);
-        }
-
-        public string FontIcon
-        {
-            get => _fontIcon;
-            set => Set(ref _fontIcon, value);
         }
 
         public bool ShowSymbolFontsOnly
@@ -222,20 +133,9 @@ namespace CharacterMap.ViewModels
                 }
 
                 RaisePropertyChanged();
-
-                if (null != _selectedFont)
-                {
-                    SelectedVariant = _selectedFont.DefaultVariant;
-                }
             }
         }
 
-        private TypographyFeatureInfo _selectedTypography;
-        public TypographyFeatureInfo SelectedTypography
-        {
-            get => _selectedTypography;
-            set => Set(ref _selectedTypography, value);
-        }
 
         private bool _isDarkAccent;
         private string _titlePrefix;
@@ -264,10 +164,7 @@ namespace CharacterMap.ViewModels
             IsLoadingFonts = false;
         }
 
-        private void LoadChars(FontVariant variant)
-        {
-            Chars = variant?.GetCharacters();
-        }
+        
 
         private void ToggleFullScreenMode()
         {
@@ -313,11 +210,11 @@ namespace CharacterMap.ViewModels
                     {
                         this.SelectedFont = lastSelectedFont;
 
-                        var lastSelectedChar = Chars.FirstOrDefault((i => i.UnicodeIndex == App.AppSettings.LastSelectedCharIndex));
-                        if (null != lastSelectedChar)
-                        {
-                            this.SelectedChar = lastSelectedChar;
-                        }
+                        //var lastSelectedChar = Chars.FirstOrDefault((i => i.UnicodeIndex == App.AppSettings.LastSelectedCharIndex));
+                        //if (null != lastSelectedChar)
+                        //{
+                        //    this.SelectedChar = lastSelectedChar;
+                        //}
                     }
                     else
                     {
@@ -335,29 +232,7 @@ namespace CharacterMap.ViewModels
             }
         }
 
-        private void UpdateCharAnalysis()
-        {
-            if (SelectedChar == null)
-            {
-                SelectedCharAnalysis = new CanvasTextLayoutAnalysis();
-                return;
-            }
-
-            using (CanvasTextLayout layout = new CanvasTextLayout(CanvasDevice.GetSharedDevice(), $"{SelectedChar.Char}", new CanvasTextFormat
-            {
-                FontSize = 20,
-                FontFamily = SelectedVariant.XamlFontFamily.Source,
-                FontStretch = SelectedVariant.FontFace.Stretch,
-                FontWeight = SelectedVariant.FontFace.Weight,
-                FontStyle = SelectedVariant.FontFace.Style,
-                HorizontalAlignment = CanvasHorizontalAlignment.Center,
-            }, 100, 100))
-            {
-                layout.Options = CanvasDrawTextOptions.EnableColorFont;
-                ApplyEffectiveTypography(layout);
-                SelectedCharAnalysis = _interop.Analyze(layout);
-            }
-        }
+       
 
         internal async void TryRemoveFont(InstalledFont font)
         {
@@ -389,43 +264,7 @@ namespace CharacterMap.ViewModels
             }
         }
 
-        private CanvasTypography GetEffectiveTypography()
-        {
-            CanvasTypography typo = new CanvasTypography();
-            if (SelectedTypography != null && SelectedTypography.Feature != CanvasTypographyFeatureName.None)
-            {
-                typo.AddFeature(SelectedTypography.Feature, 1u);
-            }
-            return typo;
-        }
-
-        private void ApplyEffectiveTypography(CanvasTextLayout layout)
-        {
-            using (var type = GetEffectiveTypography())
-            {
-                layout.SetTypography(0, 1, type);
-            }
-        }
-
-        private Task SavePngAsync(ExportStyle style)
-        {
-            return ExportManager.ExportPngAsync(
-                style,
-                SelectedFont,
-                SelectedVariant,
-                SelectedChar,
-                GetEffectiveTypography());
-        }
-
-        private Task SaveSvgAsync(bool isBlackText)
-        {
-            return ExportManager.ExportSvgAsync(
-                isBlackText ? ExportStyle.Black : ExportStyle.White,
-                SelectedFont,
-                SelectedVariant,
-                SelectedChar,
-                GetEffectiveTypography());
-        }
+        
 
     }
 }
